Revert "fix headers/body order (#55)"
This reverts commit e63e657919499bc0a826c5fce82ce451ba98ea8f.

The commit above introduced sane-ordering to the terminal output by
prioritizing headers first and then the response body. But unfortunately
it introduces a deadlock in cases where we have no headers to print.

Commands like `curlie` or `curlie -h` will hang.

Sometimes it also causes the program to deadlock even when headers are
present. It probably has something to do with the code failing to
properly detect where the headers' section ends. But this is difficult
to reproduce.